delete post

+++ OrderManagement/Controllers/HomeController.cs
@@ -24,14 +24,13 @@ namespace OrderManagement.Controllers
             }
             return View(db.Orders.Include(o => o.Customer).Include(o => o.Employee).Include(o => o.Shipper).ToList());
         }
-    
+
         public ActionResult EmployeeDetails(int id = 0)
         {
             Employee emp = db.Employees.Find(id);
             ViewBag.empName = db.Employees.Find(emp.ReportsTo).FirstName + " " + db.Employees.Find(emp.ReportsTo).LastName; //must be changed
             return PartialView("_EmployeeDetails", emp);
         }
-
 
         public ActionResult Details(int id = 0)
         {
@@ -43,9 +42,6 @@ namespace OrderManagement.Controllers
             return View(order);
         }
 
-        //
-        // GET: /Home/Create
-
         public ActionResult Create()
         {
             ViewBag.CustomerID = new SelectList(db.Customers, "CustomerID", "CompanyName");
@@ -53,9 +49,6 @@ namespace OrderManagement.Controllers
             ViewBag.ShipVia = new SelectList(db.Shippers, "ShipperID", "CompanyName");
             return View();
         }
-
-        //
-        // POST: /Home/Create
 
         [HttpPost]
         public ActionResult Create(Order order)
@@ -73,9 +66,6 @@ namespace OrderManagement.Controllers
             return View(order);
         }
 
-        //
-        // GET: /Home/Edit/5
-
         public ActionResult Edit(int id = 0)
         {
             Order order = db.Orders.Find(id);
@@ -84,13 +74,10 @@ namespace OrderManagement.Controllers
                 return HttpNotFound();
             }
             ViewBag.CustomerID = new SelectList(db.Customers, "CustomerID", "CompanyName", order.CustomerID);
-            ViewBag.EmployeeID = new SelectList(db.Employees, "EmployeeID", "LastName",order.EmployeeID);
+            ViewBag.EmployeeID = new SelectList(db.Employees, "EmployeeID", "LastName", order.EmployeeID);
             ViewBag.ShipVia = new SelectList(db.Shippers, "ShipperID", "CompanyName", order.ShipVia);
             return View(order);
         }
-
-        //
-        // POST: /Home/Edit/5
 
         [HttpPost]
         public ActionResult Edit(Order order)
@@ -107,27 +94,25 @@ namespace OrderManagement.Controllers
             return View(order);
         }
 
-        //
-        // GET: /Home/Delete/5
-
         public ActionResult Delete(int id = 0)
         {
-            var order = db.Orders.Where(o=>o.OrderID==id);
+            var order = db.Orders.Where(o => o.OrderID == id);
             order.Include(o => o.Customer);
-            
             return PartialView("_DeleteDetails", order.ToList());
         }
 
-        //
-        // POST: /Home/Delete/5
-
-        //[HttpPost, ActionName("Delete")]
+        [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
-            Order order = db.Orders.Find(id);
-            db.Orders.Remove(order);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            var order = db.Orders.Find(id);
+            if (order != null)
+            {
+                db.Orders.Remove(order);
+                db.SaveChanges();
+               
+            }
+
+            return RedirectToAction("Index");        
         }
 
         protected override void Dispose(bool disposing)
